pyproject.toml
[project]
name = "sarhelper"
version = "0.1.0"
requires-python = ">= 3.9"
dependencies = [
  "black==24.1.1",
  "flake8==7.0.0",
  "isort==5.13.2",
  "pre-commit==3.6.0",
  "hydra-core==1.3.2",
  "matplotlib==3.8.2",
  "numpy==1.26.4",
  "pytest==8.0.0",
  "rioxarray==0.15.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
proc = "proc:main"
